moved PoW related stuff into its own file
